+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useResendAudience';
 export * from './useWaitlistForm';
-export * from './useWaitlistEvents'; 
+export * from './useWaitlistEvents';
+export * from './useReCaptcha';